+++ Rpackage/iMRMC/NEWS.md
@@ -11,6 +11,12 @@ A big thank you to all that have worked on this software throughout the years. W
 *	Yousun Ko &emsp;&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;&emsp; •	Rowhan Pathare  
 * Si Wen &emsp;&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;&emsp;&nbsp;&nbsp; •	Adam Wunderlich   
 
+# iMRMC 2.1.0
+*08/2024*
+- Add functions and tests to do MRMC analysis of limits of agreement `laWRWM`, `laWRBM`, `laBRWM`, `laBRBM`
+- Add function and test to simulate MRMC agreement data: `sim.NormalIG.Hierarchical`
+- Add scripts to demonstrate new simulation and analysis of MRMC agreement data in the `Rpackage/iMRMC/inst/extra/limitsOfAgreement` folder
+
 # iMRMC 2.0.0
 *04/2024*
 - Add R only functions to analyze MRMC studies, notably `doIMRMC()`